macro: Consolidate parse layer into parse/ module
Phase 1 of the contract-macro pipeline (tokens -> IR) was fragmented
across parse.rs (use-tree only), extract.rs (kitchen sink, 1714 lines),
and lib.rs (visitors, attribute parsers, dedup). Each phase of the
pipeline (validate, resolve, generate, data_driver) had its own home;
parse did not.

Reorganize into a parse/ module with one submodule per IR-producing
concern:

  parse/
    mod.rs         re-exports + ContractData orchestrator
    module.rs      walks user mod {} body
    imports.rs     use-tree -> ImportInfo
    functions.rs   impl block -> FunctionInfo / ParameterInfo
    events.rs      abi::emit() / abi::feed() discovery -> EventInfo
    directives.rs  #[contract(...)] directive parsers (as-is)

Pure file moves; tests travel with their code. The macro entry point's
signature and observable behavior are unchanged: `make test` reports
identical test counts (157 unit + 2 trybuild + 2 doctests) and
`cargo expand` (both contract and data-driver-js features) produces
byte-identical output to main.
